+++ packages/stream_video_screen_sharing/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.5.4
+* **Flutter SDK** constraint updated to >=3.22.0 (**Dart SDK** to >=3.4.0 <4.0.0)
+
+## 0.5.3
+* Sync version with `stream_video_flutter` 0.5.3
+
+## 0.5.2
+* Sync version with `stream_video_flutter` 0.5.2
+
 ## 0.5.1
 * Sync version with `stream_video_flutter` 0.5.1